feat: improve architecture of perks rendering, also fix visual issues
