+++ src/components/badges/modals/LevelUpModal/index.tsx
@@ -1,10 +1,8 @@
-import React, { useMemo } from 'react'
-import { Box, Button, Dialog, Skeleton, SvgIcon, Typography } from '@mui/material'
+import React from 'react'
+import { Box, Button, Dialog, Typography } from '@mui/material'
 import css from './styles.module.css'
-import Tooltip from '@mui/material/Tooltip'
-import PerkRaffle from '@/public/images/superchain/perk-raffle.svg'
-import PerkSponsored from '@/public/images/superchain/perk-gas.svg'
 import StarAnimation from '../StarsAnimation'
+import Perks from '@/components/superChain/Perks'
 import { useQuery } from '@tanstack/react-query'
 import badgesService from '@/features/superChain/services/badges.service'
 function LevelUpModal({ open, onClose, level }: { open: boolean; onClose: () => void; level: number }) {
@@ -12,24 +10,6 @@ function LevelUpModal({ open, onClose, level }: { open: boolean; onClose: () => 
     queryKey: ['levelUpModal', level],
     queryFn: async () => badgesService.getPerksByLevel(level),
   })
-  const perks = useMemo(() => {
-    console.debug('data', data)
-    if (!data) {
-      return {
-        raffle: { value: 0 },
-
-        sponsoredTxns: { value: 0 },
-      }
-    }
-    return {
-      raffle: {
-        value: data.find((perk) => perk.name === 'SuperChainRaffle')?.value ?? 0,
-      },
-      sponsoredTxns: {
-        value: data.find((perk) => perk.name === 'SponsoredTxns')?.value ?? 0,
-      },
-    }
-  }, [data])
 
   return (
     <>
@@ -43,7 +23,6 @@ function LevelUpModal({ open, onClose, level }: { open: boolean; onClose: () => 
         <Box
           display="flex"
           flexDirection="column"
-          gap="24px"
           padding="36px 24px 36px 24px"
           justifyContent="center"
           alignItems="center"
@@ -56,66 +35,8 @@ function LevelUpModal({ open, onClose, level }: { open: boolean; onClose: () => 
               You have unlocked the following perks:
             </Typography>
           </Box>
-          {isLoading ? (
-            <>
-              <Skeleton variant="rounded" width="100%" height={40} />
-              <Skeleton variant="rounded" width="100%" height={40} />
-              <Skeleton variant="rounded" width="100%" height={40} />
-            </>
-          ) : (
-            <>
-              <Box display="flex" width="80%" flexDirection="column" gap="24px" flexWrap="wrap">
-                <Box
-                  display="flex"
-                  width="100%"
-                  flexDirection="row"
-                  justifyContent="flex-start"
-                  gap="12px"
-                  alignItems="center"
-                >
-                  <Tooltip title={<Typography align="center">SuperChain Raffle</Typography>}>
-                    <Box display="flex" justifyContent="center" alignItems="center">
-                      <SvgIcon component={PerkRaffle} inheritViewBox className={css.perk} />
-                    </Box>
-                  </Tooltip>
-                  <Typography
-                    fontSize={16}
-                    border={1}
-                    borderColor="secondary.main"
-                    borderRadius="6px"
-                    padding="12px"
-                    width="100%"
-                  >
-                    Claim {perks.raffle?.value ?? 0} tickets per week
-                  </Typography>
-                </Box>
-              </Box>
-              <Box
-                display="flex"
-                width="100%"
-                flexDirection="row"
-                justifyContent="flex-start"
-                gap="12px"
-                alignItems="center"
-              >
-                <Tooltip title={<Typography align="center">Sponsored Transactions</Typography>}>
-                  <Box display="flex" justifyContent="center" alignItems="center">
-                    <SvgIcon component={PerkSponsored} inheritViewBox className={css.perk} />
-                  </Box>
-                </Tooltip>
-                <Typography
-                  fontSize={16}
-                  border={1}
-                  borderColor="secondary.main"
-                  borderRadius="6px"
-                  padding="12px"
-                  width="100%"
-                >
-                  {perks.sponsoredTxns?.value ?? 0} Sponsored Transactions per week
-                </Typography>
-              </Box>
-            </>
-          )}
+
+          <Perks data={data} isLoading={isLoading} />
         </Box>
         <Button onClick={onClose} variant="contained" className={css.outsideButton}>
           Return to Dashboard
